Add send remote keys

+++ app/src/main/java/de/wollis_page/gibsonos/module/hc/module/ir/dto/remote/Key.kt
@@ -16,6 +16,6 @@ data class Key(
     val borderRadiusTopRight: Int,
     val borderRadiusBottomLeft: Int,
     val borderRadiusBottomRight: Int,
-    val eventId: Int?,
+    val eventId: Long?,
     val keys: MutableList<Key>,
 )
